% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_add_case}
\alias{qc_add_case}
\title{Add a case (respondent / subject)}
\usage{
qc_add_case(project, name, memo = "")
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{name}{Character. Case label. Must be unique.}

\item{memo}{Character.}
}
\value{
A one-row tibble: \code{id}, \code{name}, \code{memo}, \code{created_at}.
}
\description{
Add a case (respondent / subject)
}
