% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{qc_add_category}
\alias{qc_add_category}
\title{Add a code category}
\usage{
qc_add_category(project, name, memo = "")
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{name}{Character. Category name. Must be unique.}

\item{memo}{Character.}
}
\value{
A one-row tibble: \code{id}, \code{name}, \code{memo}.
}
\description{
Add a code category
}
