% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{qc_delete_annotation}
\alias{qc_delete_annotation}
\title{Delete an annotation (soft delete)}
\usage{
qc_delete_annotation(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Annotation id.}
}
\value{
Invisibly \code{1L}.
}
\description{
Delete an annotation (soft delete)
}
