% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_delete_case}
\alias{qc_delete_case}
\title{Soft-delete a case}
\usage{
qc_delete_case(project, case_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Soft-delete a case
}
