% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_delete_case_attribute}
\alias{qc_delete_case_attribute}
\title{Delete a case attribute}
\usage{
qc_delete_case_attribute(project, case_id, variable)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}

\item{variable}{Character. Attribute name to remove.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Delete a case attribute
}
