% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{qc_delete_code}
\alias{qc_delete_code}
\title{Delete a code (soft delete)}
\usage{
qc_delete_code(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Code id.}
}
\value{
Invisibly, the number of codings also soft-deleted.
}
\description{
Also soft-deletes all codings that used this code.
}
