% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_relations.R
\name{qc_delete_code_relation}
\alias{qc_delete_code_relation}
\title{Delete a code relation (soft delete)}
\usage{
qc_delete_code_relation(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Relation id.}
}
\value{
Invisibly \code{1L}.
}
\description{
Delete a code relation (soft delete)
}
