% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{qc_delete_coding}
\alias{qc_delete_coding}
\title{Delete a single coding (soft delete)}
\usage{
qc_delete_coding(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. The coding id.}
}
\value{
Invisibly, \code{1L}.
}
\description{
Delete a single coding (soft delete)
}
