% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{qc_delete_document}
\alias{qc_delete_document}
\title{Remove a document (soft delete)}
\usage{
qc_delete_document(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Document id.}
}
\value{
Invisibly, the number of codings also soft-deleted.
}
\description{
Also soft-deletes all codings attached to this document.
}
