% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerpts.R
\name{qc_delete_excerpt}
\alias{qc_delete_excerpt}
\title{Delete an excerpt (soft delete)}
\usage{
qc_delete_excerpt(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Excerpt id.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Delete an excerpt (soft delete)
}
