% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memos.R
\name{qc_delete_project_memo}
\alias{qc_delete_project_memo}
\title{Delete a project journal entry (soft delete)}
\usage{
qc_delete_project_memo(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Memo id.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Delete a project journal entry (soft delete)
}
