% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qc_delete_theme}
\alias{qc_delete_theme}
\title{Delete a theme (soft delete)}
\usage{
qc_delete_theme(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Theme id.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Delete a theme (soft delete)
}
