% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_link_case_source}
\alias{qc_link_case_source}
\title{Link a case to a document}
\usage{
qc_link_case_source(project, case_id, source_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}

\item{source_id}{Integer.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Link a case to a document
}
