% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_list_case_attributes}
\alias{qc_list_case_attributes}
\title{List attributes for one case (long format)}
\usage{
qc_list_case_attributes(project, case_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}
}
\value{
A tibble: \code{variable}, \code{value}.
}
\description{
List attributes for one case (long format)
}
