% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_list_cases}
\alias{qc_list_cases}
\title{List all cases}
\usage{
qc_list_cases(project)
}
\arguments{
\item{project}{A \code{qc_project} object.}
}
\value{
A tibble: \code{id}, \code{name}, \code{memo}, \code{n_sources}, \code{created_at}.
}
\description{
List all cases
}
