% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{qc_list_coders}
\alias{qc_list_coders}
\title{List coders and their coding activity}
\usage{
qc_list_coders(project)
}
\arguments{
\item{project}{A \code{qc_project} object.}
}
\value{
A tibble: \code{coder}, \code{n_codings}, \code{n_documents}, \code{n_codes}.
}
\description{
List coders and their coding activity
}
