% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerpts.R
\name{qc_list_excerpts}
\alias{qc_list_excerpts}
\title{List excerpts}
\usage{
qc_list_excerpts(project, source_id = NULL)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{source_id}{Integer or \code{NULL}. Filter to a single document.}
}
\value{
A tibble: \code{id}, \code{source_id}, \code{source_name}, \code{selfirst}, \code{selast},
\code{seltext}, \code{memo}, \code{coder}, \code{created_at}.
}
\description{
List excerpts
}
