% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qc_list_themes}
\alias{qc_list_themes}
\title{List all themes}
\usage{
qc_list_themes(project)
}
\arguments{
\item{project}{A \code{qc_project} object.}
}
\value{
A tibble: \code{id}, \code{name}, \code{central_concept}, \code{narrative},
\code{definition}, \code{scope}, \code{n_categories}, \code{n_codes}, \code{status}, \code{created_at}.
}
\description{
List all themes
}
