% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qc_open}
\alias{qc_open}
\title{Open an existing saturate project}
\usage{
qc_open(path, read_only = FALSE)
}
\arguments{
\item{path}{Character. Path to an existing \code{.duckdb} file.}

\item{read_only}{Logical. Open in read-only mode.}
}
\value{
A \code{qc_project} object.
}
\description{
Open an existing saturate project
}
