% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{qc_plot_saturation}
\alias{qc_plot_saturation}
\title{Plot the code saturation curve}
\usage{
qc_plot_saturation(project, ..., dark = FALSE)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{...}{Additional arguments passed to \code{\link[=qc_saturation_curve]{qc_saturation_curve()}}.}

\item{dark}{Logical. Apply dark-mode plot colours.}
}
\value{
A \code{ggplot} object.
}
\description{
Requires \code{ggplot2}.
}
