% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_set_case_attribute}
\alias{qc_set_case_attribute}
\title{Set a case attribute value (upsert)}
\usage{
qc_set_case_attribute(project, case_id, variable, value)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}

\item{variable}{Character. Attribute name (e.g. \code{"age_group"}).}

\item{value}{Character.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Set a case attribute value (upsert)
}
