% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{qc_undeprecate_code}
\alias{qc_undeprecate_code}
\title{Restore a deprecated code to active status}
\usage{
qc_undeprecate_code(project, id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Code id.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Clears the deprecated flag so new codings can be applied again.
}
