% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_unlink_case_source}
\alias{qc_unlink_case_source}
\title{Remove a case-document link}
\usage{
qc_unlink_case_source(project, case_id, source_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}

\item{source_id}{Integer.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Remove a case-document link
}
