% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{qc_unlink_code_category}
\alias{qc_unlink_code_category}
\title{Remove a code from a category}
\usage{
qc_unlink_code_category(project, code_id, category_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{code_id}{Integer.}

\item{category_id}{Integer.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Remove a code from a category
}
