% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qc_unlink_theme_category}
\alias{qc_unlink_theme_category}
\title{Unlink a category from a theme}
\usage{
qc_unlink_theme_category(project, theme_id, category_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{theme_id}{Integer. Theme id.}

\item{category_id}{Integer. Category id.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Unlink a category from a theme
}
