% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qc_unlink_theme_code}
\alias{qc_unlink_theme_code}
\title{Unlink a code from a theme}
\usage{
qc_unlink_theme_code(project, theme_id, code_id)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{theme_id}{Integer. Theme id.}

\item{code_id}{Integer. Code id.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Unlink a code from a theme
}
