% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{qc_update_annotation}
\alias{qc_update_annotation}
\title{Update annotation text}
\usage{
qc_update_annotation(project, id, annotation)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Annotation id.}

\item{annotation}{Character. New text.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Update annotation text
}
