% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{qc_update_case}
\alias{qc_update_case}
\title{Update a case name and/or memo}
\usage{
qc_update_case(project, case_id, name = NULL, memo = NULL)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{case_id}{Integer.}

\item{name}{Character. New name (or \code{NULL} to leave unchanged).}

\item{memo}{Character. New memo (or \code{NULL} to leave unchanged).}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Update a case name and/or memo
}
