% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{qc_update_coding_confidence}
\alias{qc_update_coding_confidence}
\title{Update the confidence score on a single coding}
\usage{
qc_update_coding_confidence(project, id, confidence)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Coding id.}

\item{confidence}{Integer 0–100 or \code{NULL} (unrated).}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Update the confidence score on a single coding
}
