% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{qc_update_coding_memo}
\alias{qc_update_coding_memo}
\title{Update the memo on a single coding}
\usage{
qc_update_coding_memo(project, id, memo)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Coding id.}

\item{memo}{Character. New memo text.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Update the memo on a single coding
}
