% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excerpts.R
\name{qc_update_excerpt_memo}
\alias{qc_update_excerpt_memo}
\title{Update an excerpt's memo}
\usage{
qc_update_excerpt_memo(project, id, memo)
}
\arguments{
\item{project}{A \code{qc_project} object.}

\item{id}{Integer. Excerpt id.}

\item{memo}{Character. New memo text.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Update an excerpt's memo
}
