% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturate-package.R
\docType{package}
\name{saturate-package}
\alias{saturate-package}
\alias{saturate}
\title{saturate: Qualitative Text Coding with DuckDB and Shiny}
\description{
A modern replacement for RQDA. Manage coding projects in DuckDB, annotate
text passages programmatically or via a Shiny GUI, and retrieve coded
segments as tidy tibbles.
\subsection{Core workflow}{
\enumerate{
\item Create or open a project with \code{\link[=qc_new]{qc_new()}} / \code{\link[=qc_open]{qc_open()}}.
\item Import documents with \code{\link[=qc_import_document]{qc_import_document()}}.
\item Define codes with \code{\link[=qc_add_code]{qc_add_code()}}.
\item Code passages with \code{\link[=qc_add_coding]{qc_add_coding()}} or interactively via \code{\link[=shiny_saturate]{shiny_saturate()}}.
\item Retrieve results with \code{\link[=qc_get_coded_segments]{qc_get_coded_segments()}}.
\item Close the project with \code{\link[=qc_close]{qc_close()}}.
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://woodtho.github.io/saturate}
  \item \url{https://github.com/woodtho/saturate}
  \item Report bugs at \url{https://github.com/woodtho/saturate/issues}
}

}
\author{
\strong{Maintainer}: Thomas Wood \email{thomawood1994@gmail.com}

}
